requirements.txt
httpx==0.15.5
aiokafka==0.6.0
python-dateutil==2.8.1
